Continues Integraion (#1)
* Update branch names in GitHub workflow

* Update CMake configuration and remove unnecessary script

* Update CMake configuration and add platform-specific build options

* Add clang and clang++ installation for Ubuntu

* Update build matrix and add Ubuntu-specific steps

* Clang path added

* Remove gtest/1.14.0 requirement from conanfile.txt

* Test

* Commented out unused steps in cmake-multi-platform.yml

* Update CMake configuration and Conan profile***

* Update compiler settings in setup.sh

* Update CMake workflow and dependencies

* Remove unnecessary platform specifier in CMakeLists.txt

* Add Qt6::QWindowsIntegrationPlugin to target_link_libraries for Windows

* Disable console window on Windows for GUI application

* Refactor CMake configuration and add ls commands for debugging

* Refactor test directory path in CMake workflow

* Removed LS

* Remove unnecessary ls command in workflow file

* Add BUILD_TESTS flag to CMake configuration

* Update workflow names

* Add path message for tests subdirectory

* Update build and installation process

* Update build-debug-test.yml and conanfile.txt

* Enable Free Disk Space check on Ubuntu

* Update CMake flags in build configuration

* Remove unnecessary test files

* Add CMakeLists.txt for tests

* Update build output directory path

* Fix build-output-dir path in build-debug-test.yml

* Refactor build configuration and test directory***

* Update build-debug-test.yml to activate conan environment based on OS

* Remove unnecessary comment in CMake configuration

* Refactor build scripts and presets

* Update build and test workflow

* Add build_type parameter to Windows job in build-and-test.yml

* Refactor reusable strings setup in build-and-test.yml

* Update build-and-test.yml with cmake configuration for Linux and Windows

* Fix build and test workflow***

***Added 'Tree' step for Windows***
***Updated comments for clarity***
***Fixed indentation in CMake configuration and build steps***
***Fixed indentation in test execution step

* Fix CMAKE_TOOLCHAIN_FILE path in build-and-test.yml

* Update CMAKE_TOOLCHAIN_FILE path in build-and-test.yml

* Update build-and-test.yml: Changed 'Tree' step to 'Dir' step for Windows

* Update build-and-test.yml to use PowerShell command for listing files on Windows

* Fix CMAKE_TOOLCHAIN_FILE path in build script

* Remove unnecessary 'Dir' step in build-and-test workflow

* Update CMAKE_TOOLCHAIN_FILE path in build-and-test.yml

* Fix CMAKE_TOOLCHAIN_FILE path in build-and-test.yml

* Update branch name in workflow file

+++ build.sh
@@ -3,8 +3,18 @@ if [ ! -d "build" ]; then
   mkdir build
 fi
 
-conan install . --build=missing --profile=default 
-cmake --preset vs2022-rel
-cmake --build --preset vs2022-rel
+if [ $# -eq 0 ]; then
+    echo "No arguments supplied, pass either rel or deb"
+    exit 1
+fi
 
-rm CMakeUserPresets.json
+if [ $1 == "rel" ]; then
+    conan install . --build=missing --profile=default 
+    rm CMakeUserPresets.json
+fi
+
+if [ $1 == "deb" ]; then
+    conan install . --build=missing --profile=Debug
+    rm CMakeUserPresets.json
+fi
+
